- initial work to put the Trajectory code into a separate object
