+++ Code/GraphMol/test_data/water_coords2.trc
