+++ Code/GraphMol/test_data/water_coords_bad2.trc
